Feat: 카카오 로그인 및 DB 저장 기능 추가2 카카오 로그인 시 status변경 DB저장 후 mainAcivity 적용

+++ ChatSoonE/app/src/main/java/com/chat_soon_e/chat_soon_e/data/local/UserDao.kt
@@ -11,16 +11,16 @@ interface UserDao {
 
     @Update
     fun update(user:User)
-
+    //유저의 상태 업데이트
     @Query("UPDATE UserTable SET status= :status WHERE idx= :id")
-    fun updateStatus(id:Int, status:String)
-
+    fun updateStatus(id:Long, status:String)
 
     @Delete
     fun delete(user:User)
 
+
     @Query("SELECT * FROM UserTable WHERE idx= :id")
-    fun getUser(id:Int):User?
+    fun getUser(id:Long):User?
 
     @Query("SELECT * FROM UserTable")
     fun getUsers():List<User>?
